+++ screens/BordScreen.js
@@ -9,15 +9,15 @@ import useRematchDispatch from "../hooks/useRematchDispatch";
 
 export default function BordScreen() {
     const {loadCountries} = useRematchDispatch(dispatch => ({
-      loadCountries: dispatch.appState.loadCountries,
+      loadCountries: dispatch.location.loadCountries,
     }));
-  
+
     useEffect(() => {
       loadCountries();
         }
-  
+
     );
-  
+
     return (
       <View style={C19Styles.container}>
         <ScrollView style={C19Styles.container} contentContainerStyle={C19Styles.contentContainer}>
