+++ screens/HomeScreen.js
@@ -8,7 +8,7 @@ import useRematchDispatch from "../hooks/useRematchDispatch";
 
 export default function HomeScreen() {
   const {loadCountries} = useRematchDispatch(dispatch => ({
-    loadCountries: dispatch.appState.loadCountries,
+    loadCountries: dispatch.location.loadCountries,
   }));
 
   useEffect(() => {
